Added interface MaterialProcess.java
